@@ -5,6 +5,7 @@ using System.Text;
 using System.Linq;
 using System.Management;
 using System.IO.Ports;
+using System.Text.RegularExpressions;
 
 namespace PressureMeasurementApplication.Model.SerialPort
 {
@@ -14,20 +15,18 @@ namespace PressureMeasurementApplication.Model.SerialPort
         //获取本机所有COM端口设备
         public Dictionary<string, string> GetPorts()
         {
-            using var searcher = new ManagementObjectSearcher(@"Select * From Win32_SerialPort");
-
+            using var searcher = new ManagementObjectSearcher(@"SELECT DeviceID,Caption FROM Win32_PnPEntity Where pnpclass = 'Ports' ");
             using var moc = searcher.Get();
 
             return moc.OfType<ManagementBaseObject>().ToDictionary(
-                x => x.GetPropertyValue("Description") + "(" + x.GetPropertyValue("DeviceID") + ")",
-                x => x.GetPropertyValue("DeviceID").ToString());
+                x => x.GetPropertyValue("Caption").ToString() ,
+                x => Regex.Match(x.GetPropertyValue("DeviceID").ToString(), @"^.+\((COM\d+)\)$").Groups[1].Value);
         }
 
         //获取所有可设置的波特率
-        public Dictionary<string, int> GetBaudRates()
-        {
-            return new[] { 300, 600, 1200, 2400, 4800, 9600, 19200, 38400, 57600, 115200, 230400 }.ToDictionary(x => $"{x} Baud", x => x);
-        }
+        public Dictionary<string, int> GetBaudRates() 
+            => new[] { 300, 600, 1200, 2400, 4800, 9600, 19200, 38400, 57600, 115200, 230400 }.ToDictionary(x => $"{x} Baud", x => x);
+
 
         //获取所有校验位
         public Dictionary<string, Parity> GetParities()
@@ -44,19 +43,28 @@ namespace PressureMeasurementApplication.Model.SerialPort
 
         //获取所有数据位
         public Dictionary<int, int> GetDataBits()
-        {
-            return new[] { 5, 6, 7, 8 }.ToDictionary(x => x, x => x);
-        }
+            => new[] { 5, 6, 7, 8 }.ToDictionary(x => x, x => x);
 
         //获取所有停止位
         public Dictionary<string, StopBits> GetStopBits()
         {
             return new Dictionary<string, StopBits>()
             {
-                {"无停止位", StopBits.None },
+                //{"无停止位", StopBits.None },
                 {"1位", StopBits.One },
                 {"1.5位", StopBits.OnePointFive },
                 {"2位", StopBits.Two },
+            };
+        }
+
+        public Dictionary<string, Handshake> GetHandshakes()
+        {
+            return new Dictionary<string, Handshake>()
+            {
+                {"无", Handshake.None },
+                {"RTS", Handshake.RequestToSend },
+                {"RTS与XON/XOFF", Handshake.RequestToSendXOnXOff },
+                {"XON/XOFF", Handshake.XOnXOff }
             };
         }
     }
