@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\View\Main.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "D4D49D4BB676FFB5AB110DD2AF401D7FE9FB8B93"
+﻿#pragma checksum "..\..\..\..\View\Main.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "1E0434E84F4E322BF5842838B5EA1B0B2D37C567"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -51,7 +51,7 @@ namespace PressureMeasurementApplication.View {
         #line hidden
         
         
-        #line 32 "..\..\..\..\View\Main.xaml"
+        #line 31 "..\..\..\..\View\Main.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.ComboBox baudRateComboBox;
         
@@ -59,7 +59,7 @@ namespace PressureMeasurementApplication.View {
         #line hidden
         
         
-        #line 49 "..\..\..\..\View\Main.xaml"
+        #line 46 "..\..\..\..\View\Main.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.ComboBox ParityComboBox;
         
@@ -67,7 +67,7 @@ namespace PressureMeasurementApplication.View {
         #line hidden
         
         
-        #line 63 "..\..\..\..\View\Main.xaml"
+        #line 60 "..\..\..\..\View\Main.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.ComboBox DataBitsComboBox;
         
@@ -75,7 +75,7 @@ namespace PressureMeasurementApplication.View {
         #line hidden
         
         
-        #line 75 "..\..\..\..\View\Main.xaml"
+        #line 72 "..\..\..\..\View\Main.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.ComboBox StopBitsComboBox;
         
